bongard_env/requirements.txt
gym==0.21.0
torch==1.11.0
numpy==1.21.5
Pillow==9.1.0
matplotlib==3.5.1
sklearn==0.0